+++ app/[communityId]/receive-project-updates/page.tsx
