+++ app/project/[projectId]/contact-info/page.tsx
@@ -1,15 +1,16 @@
 /* eslint-disable @next/next/no-img-element */
 import React from "react";
-import { ProjectPageLayout } from ".";
+import { ProjectPageLayout } from "../page";
 import { useOwnerStore, useProjectStore } from "@/store";
 import { Hex } from "viem";
 import { GetServerSidePropsContext, InferGetServerSidePropsType } from "next";
-import { IProjectDetails } from "@show-karma/karma-gap-sdk";
+import type { IProjectDetails } from "@show-karma/karma-gap-sdk";
 import { NextSeo } from "next-seo";
+import { ContactInfoSubscription } from "@/components/ContactInfoSubscription";
+import { Spinner } from "@/components/Utilities/Spinner";
 import { getMetadata } from "@/utilities/sdk";
 import { zeroUID } from "@/utilities/commons";
 import { defaultMetadata } from "@/utilities/meta";
-import { ImpactComponent } from "@/components/Pages/Project/Impact";
 
 export async function getServerSideProps(context: GetServerSidePropsContext) {
   const { params } = context;
@@ -32,7 +33,7 @@ export async function getServerSideProps(context: GetServerSidePropsContext) {
     },
   };
 }
-const ImpactPage = ({
+const ContactInfoPage = ({
   projectTitle,
   projectDesc,
 }: InferGetServerSidePropsType<typeof getServerSideProps>) => {
@@ -40,6 +41,10 @@ const ImpactPage = ({
     title: `Karma GAP - ${projectTitle}`,
     description: projectDesc,
   };
+  const contactsInfo = useProjectStore((state) => state.projectContactsInfo);
+  const contactInfoLoading = useProjectStore(
+    (state) => state.contactInfoLoading
+  );
 
   const isOwnerLoading = useOwnerStore((state) => state.isOwnerLoading);
   const isProjectOwnerLoading = useProjectStore(
@@ -66,7 +71,7 @@ const ImpactPage = ({
             url: image,
             alt: dynamicMetadata.title || defaultMetadata.title,
           })),
-          // // site_name: defaultMetadata.openGraph.siteName,
+          // site_name: defaultMetadata.openGraph.siteName,
         }}
         additionalLinkTags={[
           {
@@ -76,12 +81,24 @@ const ImpactPage = ({
         ]}
       />
       <div className="pt-5 pb-20">
-        <ImpactComponent />
+        {contactInfoLoading || isAuthorizationLoading ? (
+          <div className="px-4 py-4 rounded-md border border-transparent dark:bg-zinc-800  dark:border flex flex-col gap-4 items-start">
+            <h3 className="text-xl font-bold leading-6 text-gray-900 dark:text-zinc-100">
+              Loading contact info...
+            </h3>
+            <Spinner />
+          </div>
+        ) : (
+          <ContactInfoSubscription
+            contactInfo={contactsInfo?.[contactsInfo.length - 1]}
+            existingContacts={contactsInfo}
+          />
+        )}
       </div>
     </>
   );
 };
 
-ImpactPage.getLayout = ProjectPageLayout;
+ContactInfoPage.getLayout = ProjectPageLayout;
 
-export default ImpactPage;
+export default ContactInfoPage;
