+++ app/project/[projectId]/impact/page.tsx
@@ -1,16 +1,15 @@
 /* eslint-disable @next/next/no-img-element */
 import React from "react";
-import { ProjectPageLayout } from ".";
+import { ProjectPageLayout } from "../page";
 import { useOwnerStore, useProjectStore } from "@/store";
 import { Hex } from "viem";
 import { GetServerSidePropsContext, InferGetServerSidePropsType } from "next";
-import type { IProjectDetails } from "@show-karma/karma-gap-sdk";
+import { IProjectDetails } from "@show-karma/karma-gap-sdk";
 import { NextSeo } from "next-seo";
-import { ContactInfoSubscription } from "@/components/ContactInfoSubscription";
-import { Spinner } from "@/components/Utilities/Spinner";
 import { getMetadata } from "@/utilities/sdk";
 import { zeroUID } from "@/utilities/commons";
 import { defaultMetadata } from "@/utilities/meta";
+import { ImpactComponent } from "@/components/Pages/Project/Impact";
 
 export async function getServerSideProps(context: GetServerSidePropsContext) {
   const { params } = context;
@@ -33,7 +32,7 @@ export async function getServerSideProps(context: GetServerSidePropsContext) {
     },
   };
 }
-const ContactInfoPage = ({
+const ImpactPage = ({
   projectTitle,
   projectDesc,
 }: InferGetServerSidePropsType<typeof getServerSideProps>) => {
@@ -41,10 +40,6 @@ const ContactInfoPage = ({
     title: `Karma GAP - ${projectTitle}`,
     description: projectDesc,
   };
-  const contactsInfo = useProjectStore((state) => state.projectContactsInfo);
-  const contactInfoLoading = useProjectStore(
-    (state) => state.contactInfoLoading
-  );
 
   const isOwnerLoading = useOwnerStore((state) => state.isOwnerLoading);
   const isProjectOwnerLoading = useProjectStore(
@@ -71,7 +66,7 @@ const ContactInfoPage = ({
             url: image,
             alt: dynamicMetadata.title || defaultMetadata.title,
           })),
-          // site_name: defaultMetadata.openGraph.siteName,
+          // // site_name: defaultMetadata.openGraph.siteName,
         }}
         additionalLinkTags={[
           {
@@ -81,24 +76,12 @@ const ContactInfoPage = ({
         ]}
       />
       <div className="pt-5 pb-20">
-        {contactInfoLoading || isAuthorizationLoading ? (
-          <div className="px-4 py-4 rounded-md border border-transparent dark:bg-zinc-800  dark:border flex flex-col gap-4 items-start">
-            <h3 className="text-xl font-bold leading-6 text-gray-900 dark:text-zinc-100">
-              Loading contact info...
-            </h3>
-            <Spinner />
-          </div>
-        ) : (
-          <ContactInfoSubscription
-            contactInfo={contactsInfo?.[contactsInfo.length - 1]}
-            existingContacts={contactsInfo}
-          />
-        )}
+        <ImpactComponent />
       </div>
     </>
   );
 };
 
-ContactInfoPage.getLayout = ProjectPageLayout;
+ImpactPage.getLayout = ProjectPageLayout;
 
-export default ContactInfoPage;
+export default ImpactPage;
