+++ app/stats/page.tsx
