+++ package.json
@@ -48,7 +48,7 @@
     "millify": "^6.1.0",
     "mixpanel-browser": "^2.49.0",
     "moment": "^2.30.1",
-    "next": "^14.1.0",
+    "next": "^14.2.4",
     "next-remove-imports": "^1.0.12",
     "next-seo": "^6.4.0",
     "next-sitemap": "^4.2.3",
@@ -58,9 +58,9 @@
     "pluralize": "^8.0.0",
     "postcss-nesting": "^12.0.2",
     "rc-slider": "^10.6.2",
-    "react": "^18.2.0",
+    "react": "^18.3.1",
     "react-day-picker": "^8.10.0",
-    "react-dom": "^18.2.0",
+    "react-dom": "^18.3.1",
     "react-hook-form": "^7.49.3",
     "react-hot-toast": "^2.4.1",
     "react-infinite-scroll-component": "^6.1.0",
